+++ wp-content/uploads/civicrm/ext/firewall/firewall.civix.php
@@ -24,7 +24,7 @@ class CRM_Firewall_ExtensionUtil {
    *   Translated text.
    * @see ts
    */
-  public static function ts($text, $params = []) {
+  public static function ts($text, $params = []): string {
     if (!array_key_exists('domain', $params)) {
       $params['domain'] = [self::LONG_NAME, NULL];
     }